# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PezeshkHub super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/Profile/show": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "profile api get profile detail",
                "description": "profile api get profile detail",
                "operationId": "f18acee7b0143923d1a8b7d91ddbca5e",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user_id for get profile detail",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "workarea_education_id": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "fullname": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "fullname_fa": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "firstname": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "firstname_fa": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "lastname": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "lastname_fa": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "about": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "avatar_url": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "medical_id": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "national_id": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "mobile_number": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "country_code": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "phone_verified_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-27T20:29:58.770Z"
                                                },
                                                "is_mobile_verify": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "is_user_verify": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "is_staff": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-27T20:29:58.770Z"
                                                },
                                                "fcm_token": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "ws_token": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "birthday": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "active": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "language": {
                                                    "type": "string",
                                                    "example": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tag not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}